--- CSV_Files/box-office-by-country.csv
@@ -1,190 +1,0 @@
-﻿Production Countries,No. of Movies,Average Production Budget,Total Worldwide Box Office
-United States,"24,344","$37,946,691","$631,802,271,246"
-United Kingdom,"4,458","$30,686,479","$56,047,913,679"
-China,"2,825","$36,834,138","$48,612,652,703"
-France,"4,259","$22,092,334","$21,092,508,762"
-Japan,"2,000","$31,739,412","$18,467,102,568"
-Germany,"2,055","$30,083,214","$11,037,283,163"
-Republic of Korea,"1,899","$16,756,481","$9,704,263,830"
-Canada,"1,934","$20,478,697","$8,693,461,399"
-Australia,"1,058","$31,050,900","$8,264,329,058"
-India,"2,565","$14,919,521","$8,187,077,983"
-New Zealand,290,"$59,255,333","$7,413,416,339"
-Hong Kong,480,"$24,077,778","$7,332,136,410"
-Italy,"1,973","$16,524,375","$4,627,337,602"
-Spain,"1,871","$23,029,052","$4,479,269,219"
-Russian Federation,"1,338","$12,590,889","$2,822,917,524"
-Belgium,669,"$16,113,575","$2,382,887,342"
-Mexico,620,"$12,148,485","$2,139,788,773"
-Sweden,338,"$15,142,857","$1,464,274,504"
-Netherlands,574,"$18,963,043","$1,416,668,712"
-Ireland,370,"$17,753,500","$1,370,090,581"
-Brazil,585,"$14,559,868","$1,119,643,075"
-Turkey,"1,247","$10,733,333","$1,069,181,800"
-"Taiwan, Province of China",178,"$13,685,714","$952,034,056"
-Denmark,327,"$11,924,181","$857,541,542"
-Finland,181,"$13,414,875","$758,985,400"
-Czech Republic,411,"$31,883,333","$737,938,913"
-Poland,359,"$7,934,909","$715,259,903"
-Switzerland,287,"$21,182,727","$700,188,932"
-Argentina,921,"$11,075,000","$641,818,116"
-South Africa,180,"$18,395,526","$608,903,753"
-Hungary,142,"$34,924,417","$570,817,401"
-United Arab Emirates,50,"$24,957,500","$539,368,692"
-Malta,16,"$66,666,667","$520,548,489"
-Austria,209,"$11,041,667","$323,088,171"
-Panama,12,"$36,000,000","$305,439,992"
-Colombia,92,"$5,125,000","$280,918,323"
-Luxembourg,121,"$19,141,667","$280,338,420"
-Norway,233,"$5,539,286","$279,547,318"
-Tunisia,27,"$35,000,000","$249,467,149"
-Romania,217,"$11,443,667","$233,741,246"
-Thailand,146,"$7,387,500","$231,370,567"
-Morocco,39,"$36,225,000","$229,669,117"
-Chile,126,"$9,100,000","$198,263,677"
-Slovakia,169,"$36,000,000","$181,334,794"
-Uruguay,55,"$135,000,000","$169,200,779"
-Paraguay,17,"$135,000,000","$165,740,252"
-Iceland,80,"$17,850,000","$145,415,983"
-Philippines,173,$0,"$128,111,926"
-Israel,231,"$2,500,000","$127,938,263"
-Ukraine,85,"$26,000,000","$106,302,954"
-Peru,44,"$1,200,000","$96,900,282"
-Kuwait,7,"$80,000,000","$96,775,326"
-Lebanon,47,"$4,625,000","$87,382,532"
-Islamic Republic of Iran,105,"$3,484,286","$85,876,861"
-Greece,101,"$2,607,667","$84,287,917"
-Cuba,45,$0,"$67,634,816"
-Portugal,364,"$800,000","$65,206,176"
-Bolivarian Republic of Venezuela,28,"$1,400,000","$56,326,599"
-Bahamas,8,"$31,500,000","$56,207,252"
-Latvia,49,"$9,500,000","$53,699,230"
-Malaysia,43,"$25,000,000","$53,695,818"
-Algeria,17,"$10,100,000","$47,510,533"
-Greenland,5,"$21,000,000","$46,084,208"
-The Former Yugoslav Republic of Macedonia,36,"$5,000,000","$45,348,525"
-Bulgaria,123,"$10,750,000","$43,927,131"
-Botswana,4,$0,"$43,923,227"
-Singapore,39,$0,"$43,494,364"
-Croatia,70,"$40,000,000","$37,912,903"
-Namibia,4,"$75,000,000","$34,034,218"
-Estonia,45,"$1,250,000","$33,273,851"
-German Democratic Republic,15,"$5,000,000","$29,573,980"
-Kazakhstan,34,"$29,000,000","$28,819,388"
-Lithuania,125,"$500,000","$27,360,005"
-Cayman Islands,3,"$9,000,000","$19,894,696"
-Costa Rica,10,"$20,000,000","$19,513,972"
-Indonesia,75,"$950,000","$18,894,993"
-Serbia,61,$0,"$18,883,910"
-Saudi Arabia,22,"$85,000,000","$18,049,536"
-Pakistan,68,"$1,100,000","$16,413,610"
-Qatar,68,"$5,428,333","$15,424,276"
-Bosnia and Herzegovina,30,"$20,500,000","$14,943,578"
-Mongolia,19,"$30,000,000","$14,692,708"
-Occupied Palestinian Territory,31,$0,"$12,813,385"
-Dominican Republic,34,"$362,500","$11,545,683"
-Slovenia,54,"$1,653,333","$10,963,049"
-Egypt,68,"$2,625,000","$10,903,835"
-Viet Nam,45,"$1,000,000","$8,913,136"
-Cyprus,17,"$3,650,000","$7,865,740"
-Montenegro,5,$0,"$7,833,027"
-Belarus,15,$0,"$7,299,428"
-Mauritania,2,"$2,500,000","$6,817,760"
-Senegal,20,$0,"$6,642,683"
-Democratic People's Republic of Korea,12,$0,"$6,507,674"
-Monaco,7,"$28,000,000","$6,056,333"
-Armenia,16,$0,"$5,233,261"
-Bermuda,3,$0,"$5,200,000"
-Afghanistan,21,"$500,000","$5,067,922"
-Puerto Rico,21,"$1,000,000","$4,652,889"
-Georgia,33,"$12,000,000","$4,368,627"
-Kenya,19,"$15,000","$4,118,104"
-USSR,7,"$25,000,000","$3,085,887"
-Cameroon,9,$0,"$3,074,051"
-Republic of Kosovo,12,$0,"$3,014,936"
-Yemen,5,$0,"$2,466,867"
-Cambodia,24,"$3,000,000","$2,447,747"
-Isle of Man,3,$0,"$2,393,476"
-Papua New Guinea,3,$0,"$2,287,136"
-Northern Ireland,4,$0,"$1,979,596"
-Nepal,31,$0,"$1,893,678"
-Tajikistan,5,$0,"$1,776,779"
-Albania,13,$0,"$1,668,679"
-Myanmar,6,$0,"$1,625,005"
-Libyan Arab Jamahiriya,3,"$35,000,000","$1,522,008"
-Jordan,10,"$2,767,500","$1,381,790"
-Bhutan,4,$0,"$1,178,662"
-Guatemala,9,$0,"$1,069,363"
-Maldives,2,$0,"$993,251"
-Brunei Darussalam,1,$0,"$993,251"
-Djibouti,1,$0,"$993,251"
-Nigeria,88,"$2,616,667","$865,925"
-Lao People’s Democratic Republic,4,$0,"$856,821"
-Vanuatu,2,$0,"$574,050"
-Bangladesh,15,$0,"$475,338"
-French Polynesia,1,$0,"$443,291"
-Palau,1,$0,"$443,291"
-Ghana,11,$0,"$406,799"
-Netherlands Antilles,1,$0,"$359,849"
-Sri Lanka,7,$0,"$355,709"
-Iraq,15,$0,"$304,301"
-Chad,3,$0,"$279,502"
-Cape Verde,4,$0,"$273,078"
-Uzbekistan,4,$0,"$270,968"
-Rwanda,4,$0,"$261,759"
-Syrian Arab Republic,13,$0,"$212,604"
-Plurinational State of Bolivia,11,$0,"$206,928"
-British Indian Ocean Territory,1,$0,"$204,784"
-Cook Islands,1,$0,"$186,978"
-Sudan,6,$0,"$168,347"
-Somalia,3,$0,"$157,071"
-Mauritius,1,$0,"$135,802"
-Ethiopia,10,$0,"$113,927"
-Azerbaijan,10,"$29,120","$102,961"
-Holy See (Vatican City State),1,$0,"$85,479"
-The Democratic Republic of the Congo,5,$0,"$84,087"
-Jamaica,7,"$750,000","$82,978"
-Angola,7,$0,"$78,698"
-Republic of Moldova,4,$0,"$64,211"
-Uganda,7,$0,"$61,510"
-Côte D'Ivoire,2,$0,"$49,967"
-Haiti,7,$0,"$41,873"
-Czechoslovakia,5,$0,"$35,541"
-United Republic of Tanzania,9,"$800,000","$27,694"
-Zimbabwe,1,$0,"$25,038"
-Macao,2,$0,"$25,021"
-Congo,1,$0,"$24,139"
-Gabon,1,$0,"$23,866"
-Bahrain,1,$0,"$19,130"
-Burkina Faso,3,$0,"$7,849"
-Andorra,1,$0,"$4,325"
-Samoa,1,$0,"$4,266"
-Lesotho,2,$0,"$3,355"
-Ecuador,9,"$800,000","$2,742"
-Mozambique,2,$0,"$2,467"
-Mali,2,$0,"$1,939"
-Kyrgyzstan,7,"$1,400,000","$1,238"
-Barbados,1,$0,"$1,105"
-Antarctica,2,$0,$976
-Nicaragua,4,$0,$0
-Liechtenstein,4,$0,$0
-Trinidad and Tobago,4,$0,$0
-Saint Kitts and Nevis,2,$0,$0
-El Salvador,2,$0,$0
-Madagascar,1,$0,$0
-Niger,1,$0,$0
-Malawi,1,$0,$0
-Gambia,1,$0,$0
-Central African Republic,1,$0,$0
-Federated States of Micronesia,1,$0,$0
-Liberia,1,$0,$0
-Zambia,1,$0,$0
-Honduras,1,$0,$0
-Fiji,1,$0,$0
-Solomon Islands,1,$0,$0
-Guinea-Bissau,1,$0,$0
-Oman,1,$0,$0
-Benin,1,$0,$0
-Kiribati,1,$0,$0
-Sierra Leone,1,$0,$0
-Belize,1,$0,$0
