--- CSV_Files/human-development-index-2021.csv
@@ -1,196 +1,0 @@
-﻿country,Life expectancy at birth,Mean years of schooling,Income per capita,Gender Inequality Index,
-Afghanistan,62.0,3.0,"1,824",0.678,
-Albania,76.5,11.3,"14,131",0.144,
-Algeria,76.4,8.1,"10,800",0.499,
-Andorra,80.4,10.6,"51,167",..,
-Angola,61.6,5.4,"5,466",0.537,
-Antigua and Barbuda,78.5,9.3,"16,792",..,
-Argentina,75.4,11.1,"20,925",0.287,
-Armenia,72.0,11.3,"13,158",0.216,
-Australia,84.5,12.7,"49,238",0.073,
-Austria,81.6,12.3,"53,619",0.053,
-Azerbaijan,69.4,10.5,"14,257",0.294,
-Bahamas,71.6,12.6,"30,486",0.329,
-Bahrain,78.8,11.0,"39,497",0.181,
-Bangladesh,72.4,7.4,"5,472",0.530,
-Barbados,77.6,9.9,"12,306",0.268,
-Belarus,72.4,12.1,"18,849",0.104,
-Belgium,81.9,12.4,"52,293",0.048,
-Belize,70.5,8.8,"6,309",0.364,
-Benin,59.8,4.3,"3,409",0.602,
-Bhutan,71.8,5.2,"9,438",0.415,
-Bolivia ,63.6,9.8,"8,111",0.418,
-Bosnia and Herzegovina,75.3,10.5,"15,242",0.136,
-Botswana,61.1,10.3,"16,198",0.468,
-Brazil,72.8,8.1,"14,370",0.390,
-Brunei Darussalam,74.6,9.2,"64,490",0.259,
-Bulgaria,71.8,11.4,"23,079",0.210,
-Burkina Faso,59.3,2.1,"2,118",0.621,
-Burundi,61.7,3.1,732,0.505,
-Cabo Verde,74.1,6.3,"6,230",0.349,
-Cambodia,69.6,5.1,"4,079",0.461,
-Cameroon,60.3,6.2,"3,621",0.565,
-Canada,82.7,13.8,"46,808",0.069,
-Central African Republic,53.9,4.3,966,0.672,
-Chad,52.5,2.6,"1,364",0.652,
-Chile,78.9,10.9,"24,563",0.187,
-China,78.2,7.6,"17,504",0.192,
-Colombia,72.8,8.9,"14,384",0.424,
-Comoros,63.4,5.1,"3,142",..,
-Costa Rica,77.0,8.8,"19,974",0.256,
-Cote d'Ivoire,58.6,5.2,"5,217",0.613,
-Croatia,77.6,12.2,"30,132",0.093,
-Cuba,73.7,12.5,"7,879",0.303,
-Cyprus,81.2,12.4,"38,188",0.123,
-Czechia,77.7,12.9,"38,745",0.120,
-Democratic Republic of the Congo,59.2,7.0,"1,076",0.601,
-Denmark,81.4,13.0,"60,365",0.013,
-Djibouti,62.3,4.1,"5,025",..,
-Dominica,72.8,8.1,"11,488",..,
-Dominican Republic,72.6,9.3,"17,990",0.429,
-Ecuador,73.7,8.8,"10,312",0.362,
-Egypt,70.2,9.6,"11,732",0.443,
-El Salvador,70.7,7.2,"8,296",0.376,
-Equatorial Guinea,60.6,5.9,"12,074",..,
-Eritrea,66.5,4.9,"1,729",..,
-Estonia,77.1,13.5,"38,048",0.100,
-Eswatini,57.1,5.6,"7,679",0.540,
-Ethiopia,65.0,3.2,"2,361",0.520,
-Fiji,67.1,10.9,"9,980",0.318,
-Finland,82.0,12.9,"49,452",0.033,
-France,82.5,11.6,"45,937",0.083,
-Gabon,65.8,9.4,"13,367",0.541,
-Gambia,62.1,4.6,"2,172",0.611,
-Georgia,71.7,12.8,"14,664",0.280,
-Germany,80.6,14.1,"54,534",0.073,
-Ghana,63.8,8.3,"5,745",0.529,
-Greece,80.1,11.4,"29,002",0.119,
-Grenada,74.9,9.0,"13,484",..,
-Guatemala,69.2,5.7,"8,723",0.481,
-Guinea,58.9,2.2,"2,481",0.621,
-Guinea-Bissau,59.7,3.6,"1,908",0.627,
-Guyana,65.7,8.6,"22,465",0.454,
-Haiti,63.2,5.6,"2,848",0.635,
-Honduras,70.1,7.1,"5,298",0.431,
-Hong Kong SAR,85.5,12.2,"62,607",..,
-Hungary,74.5,12.2,"32,789",0.221,
-Iceland,82.7,13.8,"55,782",0.043,
-India,67.2,6.7,"6,590",0.490,
-Indonesia,67.6,8.6,"11,466",0.444,
-Iran,73.9,10.6,"13,001",0.459,
-Iraq,70.4,7.9,"9,977",0.558,
-Ireland,82.0,11.6,"76,169",0.074,
-Israel,82.3,13.3,"41,524",0.083,
-Italy,82.9,10.7,"42,840",0.056,
-Jamaica,70.5,9.2,"8,834",0.335,
-Japan,84.8,13.4,"42,274",0.083,
-Jordan,74.3,10.4,"9,924",0.471,
-Kazakhstan,69.4,12.3,"23,943",0.161,
-Kenya,61.4,6.7,"4,474",0.506,
-Kiribati,67.4,8.0,"4,063",..,
-North Korea ,73.3,..,..,..,
-South Korea,83.7,12.5,"44,501",0.067,
-Kuwait,78.7,7.3,"52,920",0.305,
-Kyrgyzstan,70.0,11.4,"4,566",0.370,
-Laos,68.1,5.4,"7,700",0.478,
-Latvia,73.6,13.3,"32,803",0.151,
-Lebanon,75.0,8.7,"9,526",0.432,
-Lesotho,53.1,6.0,"2,700",0.557,
-Liberia,60.7,5.1,"1,289",0.648,
-Libya,71.9,7.6,"15,336",0.259,
-Liechtenstein,83.3,12.5,"146,830",..,
-Lithuania,73.7,13.5,"37,931",0.105,
-Luxembourg,82.6,13.0,"84,649",0.044,
-Madagascar,64.5,5.1,"1,484",0.556,
-Malawi,62.9,4.5,"1,466",0.554,
-Malaysia,74.9,10.6,"26,658",0.228,
-Maldives,79.9,7.3,"15,448",0.348,
-Mali,58.9,2.3,"2,133",0.613,
-Malta,83.8,12.2,"38,884",0.167,
-Marshall Islands,65.3,10.9,"4,620",..,
-Mauritania,64.4,4.9,"5,075",0.632,
-Mauritius,73.6,10.4,"22,025",0.347,
-Mexico,70.2,9.2,"17,896",0.309,
-Micronesia,70.7,7.8,"3,696",..,
-Moldova,68.8,11.8,"14,875",0.205,
-Monaco,85.9,..,..,..,
-Mongolia,71.0,9.4,"10,588",0.313,
-Montenegro,76.3,12.2,"20,839",0.119,
-Morocco,74.0,5.9,"7,303",0.425,
-Mozambique,59.3,3.2,"1,198",0.537,
-Myanmar,65.7,6.4,"3,851",0.498,
-Namibia,59.3,7.2,"8,634",0.445,
-Nauru,63.6,..,"17,730",..,
-Nepal,68.4,5.1,"3,877",0.452,
-Netherlands,81.7,12.6,"55,979",0.025,
-New Zealand,82.5,12.9,"44,057",0.088,
-Nicaragua,73.8,7.1,"5,625",0.424,
-Niger,61.6,2.1,"1,240",0.611,
-Nigeria,52.7,7.2,"4,790",0.680,
-North Macedonia,73.8,10.2,"15,918",0.134,
-Norway,83.2,13.0,"64,660",0.016,
-Oman,72.5,11.7,"27,054",0.300,
-Pakistan,66.1,4.5,"4,624",0.534,
-Palau,66.0,12.5,"13,819",..,
-West Bank and Gaza,73.5,9.9,"6,583",..,
-Panama,76.2,10.5,"26,957",0.392,
-Papua New Guinea,65.4,4.7,"4,009",0.725,
-Paraguay,70.3,8.9,"12,349",0.445,
-Peru,72.4,9.9,"12,246",0.380,
-Philippines,69.3,9.0,"8,920",0.419,
-Poland,76.5,13.2,"33,034",0.109,
-Portugal,81.0,9.6,"33,155",0.067,
-Qatar,79.3,10.0,"87,134",0.220,
-Republic of Congo,63.5,6.2,"2,889",0.564,
-Romania,74.2,11.3,"30,027",0.282,
-Russia,69.4,12.8,"27,166",0.203,
-Rwanda,66.1,4.4,"2,210",0.388,
-Saint Kitts and Nevis,71.7,8.7,"23,358",..,
-Saint Lucia,71.1,8.5,"12,048",0.381,
-Saint Vincent and the Grenadines,69.6,10.8,"11,961",0.390,
-Samoa,72.8,11.4,"5,308",0.418,
-San Marino,80.9,10.8,"52,654",..,
-Sao Tome and Principe,67.6,6.2,"4,021",0.494,
-Saudi Arabia,76.9,11.3,"46,112",0.247,
-Senegal,67.1,2.9,"3,344",0.530,
-Serbia,74.2,11.4,"19,123",0.131,
-Seychelles,71.3,10.3,"25,831",..,
-Sierra Leone,60.1,4.6,"1,622",0.633,
-Singapore,82.8,11.9,"90,919",0.040,
-Slovakia,74.9,12.9,"30,690",0.180,
-Slovenia,80.7,12.8,"39,746",0.071,
-Solomon Islands,70.3,5.7,"2,482",..,
-Somalia,55.3,..,"1,018",..,
-South Africa,62.3,11.4,"12,948",0.405,
-South Sudan,55.0,5.7,768,0.587,
-Spain,83.0,10.6,"38,354",0.057,
-Sri Lanka,76.4,10.8,"12,578",0.383,
-Sudan,65.3,3.8,"3,575",0.553,
-Suriname,70.3,9.8,"12,672",0.427,
-Sweden,83.0,12.6,"54,489",0.023,
-Switzerland,84.0,13.9,"66,933",0.018,
-Syria,72.1,5.1,"4,192",0.477,
-Tajikistan,71.6,11.3,"4,548",0.285,
-Tanzania ,66.2,6.4,"2,664",0.560,
-Thailand,78.7,8.7,"17,030",0.333,
-Timor-Leste,67.7,5.4,"4,461",0.378,
-Togo,61.6,5.0,"2,167",0.580,
-Tonga,71.0,11.4,"6,822",0.631,
-Trinidad and Tobago,73.0,11.6,"23,392",0.344,
-Tunisia,73.8,7.4,"10,258",0.259,
-Turkey,76.0,8.6,"31,033",0.272,
-Turkmenistan,69.3,11.3,"13,021",0.177,
-Tuvalu,64.5,10.6,"6,351",..,
-Uganda,62.7,5.7,"2,181",0.530,
-Ukraine,71.6,11.1,"13,256",0.200,
-United Arab Emirates,78.7,12.7,"62,574",0.049,
-United Kingdom,80.7,13.4,"45,225",0.098,
-United States,77.2,13.7,"64,765",0.179,
-Uruguay,75.4,9.0,"21,269",0.235,
-Uzbekistan,70.9,11.9,"7,917",0.227,
-Vanuatu,70.4,7.1,"3,085",..,
-Venezuela,70.6,11.1,"4,811",0.492,
-Vietnam,73.6,8.4,"7,867",0.296,
-Yemen,63.8,3.2,"1,314",0.820,
-Zambia,61.2,7.2,"3,218",0.540,
-Zimbabwe,59.3,8.7,"3,810",0.532,
