--- CSV_Files/pet-ownership-statistics-by-country-2024.csv
@@ -1,40 +1,0 @@
-"country","PetOwnershipDogPopulation","PetOwnershipCatPopulation","PetOwnershipBirdPopulation","PetOwnershipAquariaPopulation","PetOwnershipSmallMammalPopulation","PetOwnershipTerrariaPopulation"
-"India",10200000,,,,,
-"China",27400000,53100000,,,,
-"United States",69929000,74059000,8300000,,,
-"Brazil",,,191001000,,,
-"Russia",,23150000,3280000,3750000,6601000,1076000
-"Ethiopia",5000000,,,,,
-"Japan",12000000,7300000,,,,
-"Philippines",11600000,,,,,
-"Turkey",1386000,4660000,11291000,850000,808000,157000
-"Germany",10600000,15200000,4900000,2300000,4900000,1300000
-"United Kingdom",13000000,12000000,1600000,8000000,,2200000
-"France",7600000,14900000,5800000,1252000,2800000,2100000
-"South Africa",7400000,2000000,,,,
-"Italy",8755000,10228000,12882000,1500000,1816000,1366000
-"Spain",9313000,5859000,6991000,691000,1523000,1464000
-"Argentina",9200000,3000000,,,,
-"Poland",8019000,7125000,1144000,380000,1217000,215000
-"Ukraine",4984000,7630000,784000,450000,899000,903000
-"Australia",3700000,2200000,7800000,,,
-"Romania",,4371000,292000,150000,93000,32000
-"Netherlands",1500000,1900000,230000,339000,346000,59000
-"Belgium",2029000,2536000,331000,320000,1026000,160000
-"Sweden",1044000,1703000,309000,260000,494000,60000
-"Czech Republic",2211000,1364000,69000,80000,,5000
-"Greece",657000,608000,974000,90000,417000,6000
-"Portugal",2605000,1795000,641000,83000,218000,39000
-"Hungary",2172000,2388000,412000,90000,398000,60000
-"Austria",766000,0,195000,310000,"508 000\r",129000
-"Switzerland",554000,1793000,304000,380000,655000,27000
-"Bulgaria",750000,815000,45000,70000,60000,14000
-"Denmark",617000,669000,202000,280000,373000,110000
-"Slovakia",,540000,32000,51000,46000,4000
-"Finland",781000,982000,56000,40000,157000,29000
-"Norway",494000,783000,195000,125000,169000,94000
-"Ireland",,355000,92000,27000,162000,20000
-"Lithuania",550000,605000,,,,
-"Slovenia",295000,455000,40000,35000,85000,3500
-"Latvia",270000,410000,,,,
-"Estonia",235000,290000,,,,
