--- indicators-page/index.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width">
-    <title>Who is Winning</title>
-    <link href="styles.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300&display=swap" rel="stylesheet">
-    <script src="https://d3js.org/d3.v5.min.js"></script> <!-- Add D3.js library -->
-    <script defer src="quiz.js"></script>
-</head>
-<body>
-
-    <h1>Who's Winning?</h1>
-    <h2>"USA this", "China that" - the news is covered with journalists comparing countries, but who's actually winning? Better still, who's winning in the metrics that matter to you? Scroll below for the definitive ranking of how countries compare on metrics you actually care about.</h2>
-
-        <div id="variable-dropdown">
-    <label for="indicator-select">Select an indicator:</label>
-    <select id="indicator-select">
-        <!-- Options will be populated here by JavaScript -->
-    </select>
-
- <div class="quiz-container">
-        <h3>Quiz: Select 5 countries you think will be in the top 10</h3>
-        <div class="guess-container">
-            <div class="guess" id="guess1">
-                Guess 1: <select class="quiz-select"></select>
-            </div>
-            <div class="guess" id="guess2">
-                Guess 2: <select class="quiz-select"></select>
-            </div>
-            <div class="guess" id="guess3">
-                Guess 3: <select class="quiz-select"></select>
-            </div>
-            <div class="guess" id="guess4">
-                Guess 4: <select class="quiz-select"></select>
-            </div>
-            <div class="guess" id="guess5">
-                Guess 5: <select class="quiz-select"></select>
-            </div>
-        </div>
-    </div>
-            <button class="button" onclick="checkQuiz()">Check Quiz</button>
-            <p id="quiz-result"></p>
-        </div>
-    </div>
-
-    <h2 id="chart-header">'Top 10 Countries - GDP per capita (USD)'</h2>
-    <div id="chart-container"></div>
-
-    <br>
-
-    <div id="content">
-        <div id="countryInfo"></div>
-        <div id="map"></div>
-    </div>
-    <script src="bar-graph.js"></script>
-    <br>
-    <a href='../index.html'>Home</a>
-    </div>
-    <div id="map-container"></div>
-    <div class="tooltip" style="opacity:0;"></div>
-    <script src="dropdown.js"></script>
-    <script src="map.js"></script>
-
-    <a href='https://forms.office.com/Pages/ResponsePage.aspx?id=9CL6b2hFBUGtQy461HJpV6PAlu8QsDNDvFA_VHvxVJZURDdTTlFUOTFZSTBDMEg0VFVDNjlGRUFYRC4u'target="_blank">Want to give feedback? Click here</a>
-</body>
-</html>
