--- version2/Chenged/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width">
-    <title>About us</title>
-    <link href="styles.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300&display=swap" rel="stylesheet">
-    <script src="https://d3js.org/d3.v5.min.js"></script> <!-- Add D3.js library -->
-</head>
-    <body>
-        <h1>Uh,oh - you've been Cheng'ed!</h1>
-
-        <img class="image-class" src="Cheng.jpg" alt="Chenged">
-
-    <a href='../index.html'>Home</a>
-    </body>
-</html>
