--- version2/Chenged/styles.css
@@ -1,32 +1,0 @@
-/* styles.css */
-
-/* Set background */
-html {
-  background-color: #264653;
-}
-
-/* styles for the body of the site */
-body {
-  font-family: 'Lato', sans-serif;
-  width: 80%;
-  background-color: #E9C46A;
-  border: 10px solid #E76F51;
-  padding: 20px 20px 20px 20px;
-  margin-top: auto;
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: auto;
-  text-align: center;
-}
-
-/* styles for headline */
-h1 {
-  font-family: 'Lato', sans-serif;
-  color: black;
-}
-
-img {
-    max-width: 400px; /* Adjust the maximum width of the images */
-    align-items: center;
-}
-
