--- version2/styles.css
@@ -1,277 +1,0 @@
-/* styles.css */
-
-/* Set background */
-html {
-  background-color: #264653;
-}
-
-/* styles for the body of the site */
-body {
-  font-family: 'Lato', sans-serif;
-  width: 80%;
-  background-color: #E9C46A;
-  border: 10px solid #E76F51;
-  padding: 20px 20px 20px 20px;
-  margin-top: auto;
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: auto;
-  text-align: center;
-}
-
-/* styles for headline */
-h1 {
-  font-family: 'Lato', sans-serif;
-  color: black;
-}
-
-/* styles for subheadline */
-h2,h3 {
-  font-family: 'Lato', sans-serif;
-  color: black;
-  font-size: 20px;
-}
-
-/* Center the quiz result */
-#quiz-result {
-  font-family: 'Lato', sans-serif;
-  text-align: center;
-  margin: auto;
-  width: 50%; /* Adjust the width as needed */
-  padding: 10px;
-  margin-top: 20px;
-}
-
-/* styles for paragraph */
-p {
-    color: black;
-    text-align: justify;
-}
-
-select#yVariableSelect {
-  font-family: 'Lato', sans-serif;
-  width: auto; /* Set a specific width */
-  padding: 5px; /* Add padding for better visibility and clickability */
-  font-size: 12px; /* Adjust font size */
-  margin-top: 10px; /* Adjust margin as needed */
-  font-weight:normal;
-}
-
-#chart-container-wrapper {
-  display: none;
-  font-size:20px;
-}
-
-
-#chart-container {
-  font-family: 'Lato', sans-serif;
-/*  display: none;*/
-  justify-content: center;
-  align-items: center;
-  margin-top: 20px; /* Adjust as needed */
-  font-size: 12px;
-  font-family: 'Lato', sans-serif;
-}
-
-#chart-header {
-  font-family: 'Lato', sans-serif;
-  color: black; /* Change the color as needed */
-  font-size: 24px; /* Adjust the font size */
-  font-weight: bold; /* Adjust the font weight */
-  text-decoration:underline;
-}
-
-#chart-subheader {
-  font-family: 'Lato', sans-serif;
-  color: black; /* Change the color as needed */
-  font-size: 16px; /* Adjust the font size */
-  font-style: italic; /* Add italic style or remove if not needed */
-}
-
-/* Add a new class for the chart SVG element */
-.chart-svg {
-  /* Adjust width and height as needed */
-  width: auto; /* Use a percentage to make it responsive or set a fixed width */
-  height: auto; /* Allow the height to adjust automatically */
-  text-align: center;
-  align-items: center;
-}
-
-.bar {
-  fill: #264653; /* Set the fill color using CSS */
-}
-
-.button-container {
-  display: flex;
-  justify-content: center; /* Adjust as needed */
-  margin-top: 10px;
-  gap: 10px;
-  align-items: center;
-}
-
-.button-1,
-.button-2 {
-  margin: 0;
-  margin-top: 10px;
-  background-color: #E76F51;
-  box-sizing: border-box;
-  color: #000000;
-  display: flex;
-  font-family: 'Lato', sans-serif;
-  font-size: 1rem;
-  font-weight: 700;
-  justify-content: center;
-  line-height: 1.75rem;
-  padding: .75rem 1.65rem;
-  position: relative;
-  text-align: center;
-  text-decoration: none #000000 solid;
-  text-decoration-thickness: auto;
-  width: 150px;
-  max-width: none;
-  cursor: pointer;
-  user-select: none;
-
-  -webkit-user-select: none;
-  touch-action: manipulation;
-}
-
-.button-1,
-.button-2:focus {
-  outline: 0;
-}
-
-.button-1:hover,
-.button-2:hover {
-  background-color: #F4A261; /* Change to the desired hover background color */
-  color: #000000;
-}
-
-@media (min-width: 768px) {
-  .button-1,
-  .button-2 {
-    padding: .25rem .15rem;
-    font-size: 1.0rem;
-  }
-}
-
-/* Container for guess options */
-#quiz-container {
-  font-family: 'Lato', sans-serif;
-  display: flex;
-  flex-wrap: wrap; /* Enable wrapping to the next line if needed */
-  justify-content: space-around; /* Adjust as needed */
-  align-items: flex-start; /* Align items at the top of the container */
-}
-
-/* Individual guess option box */
-#quiz-container > div {
-  font-family: 'Lato', sans-serif;
-  flex: 0 0 30%; /* Set a fixed width for each box (adjust as needed) */
-  margin: 10px; /* Adjust margin as needed */
-  padding: 10px;
-  border: 2px solid #E76F51;
-  background-color: #F4A261;
-  text-align: center;
-}
-
-/* Header for each guess option box */
-#quiz-container > div h3 {
-  font-family: 'Lato', sans-serif;
-  color: black;
-  font-size: 16px; /* Adjust font size */
-  text-align: center;
-}
-
-/* Styles for the select dropdown within each box */
-#quiz-container .quiz-select {
-  font-family: 'Lato', sans-serif;
-  height: 50px;
-  width: 75%; /* Take full width of the box */
-  margin-top: 5px; /* Adjust margin as needed */
-  text-align: center;
-}
-
-/* Center text in quiz options */
-#quiz-container .quiz-select option {
-  text-align: center;
-}
-
-/* Override styles for the <p> section within #quiz-result */
-#quiz-result p {
-  text-align: center;
-  font-family: 'Lato', sans-serif;
-  height: 50px;
-  width: 75%; /* Take full width of the box */
-  margin-top: 5px; /* Adjust margin as needed */
-  text-align: center;
-  background-color:none;
-}
-
-
-.image-class {
-  font-family: 'Lato', sans-serif;
-  max-width: 20%; /* Adjust the width as needed */
-  height: auto; /* Maintain the aspect ratio */
-}
-
-#content {
-    font-family: 'Lato', sans-serif;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 2000px;
-    margin: auto;
-}
-
-.y-axis text {
-    font-family: 'Lato', sans-serif;
-    font-size: 15px/* Your Y-axis styling here */
-}
-
-.x-axis text {
-    font-family: 'Lato', sans-serif;
-    font-size: 15px/* Your X-axis styling here */
-}
-
-#countryInfo {
-    width: 200px;
-    height: 70px; /* Adjusted height to match the map */
-    border: 3px solid darkblue;
-    padding: 10px;
-    box-sizing: border-box;
-    margin-right: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-#map-header {
-  display: none;
-}
-
-#map {
-    width: 760px;
-    height: 400px; /* Reduced height to 400px */
-    border: 3px solid darkblue;
-}
-
-.tooltip { 
-    position: absolute;           
-    text-align: center;           
-    width: 200px;                  
-    height: auto;                 
-    padding: 2px;             
-    font: 12px sans-serif;        
-    background: lightsteelblue;   
-    border: 0px;      
-    border-radius: 8px;           
-    pointer-events: none;         
-}
-
-#refreshButton {
-    margin-top: 10px;
-    margin-bottom:10px;
-    margin: 0 auto; /* Center the button horizontally */
-    display: none; /* Make sure it's a block-level element */
-}
